chore(storybook): add support for playgrounds
